Additional fixes for warning mismatch with scons.
Fix broken SDK build due to mismatch in warning level which permits
this conversions on the GYP build.

TBR=dschuff@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10532147

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@142178 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5aeb62af9303b28ca2962895bc41dbfd764fa4df

+++ ppapi/native_client/src/trusted/plugin/pnacl_streaming_translate_thread.cc
@@ -62,7 +62,7 @@ void PnaclStreamingTranslateThread::PutBytes(std::vector<char>* bytes,
                                              int count) {
   PLUGIN_PRINTF(("PutBytes, this %p bytes %p, size %d, count %d\n", this, bytes,
                  bytes ? bytes->size(): 0, count));
-  int buffer_size = 0;
+  size_t buffer_size = 0;
   // Ensure that the buffer we send to the translation thread is the right size
   // (count can be < the buffer size). This can be done without the lock.
   if (bytes != NULL && count >= 0) {
